build local storage on windows, make grace os independent
